pyproject.toml
[project]
name = "catvatar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "uvicorn>=0.40.0",
]
